+++ src/controller_device_recv.h
@@ -55,6 +55,8 @@ int device_state_recv_get_schema(device_handle dh, cxobj *xmsg, char *rpcname,
                                  conn_state conn_state);
 int device_state_recv_ok(clixon_handle h, device_handle dh, cxobj *xmsg, char *rpcname,
                          conn_state conn_state, cbuf **cberr);
+int device_state_recv_rpc(clixon_handle h, device_handle dh, controller_transaction *ct, cxobj *xmsg,
+                          char *rpcname, conn_state conn_state, cbuf **cberr);
 
 #ifdef __cplusplus
 }