+++ src/controller_transaction.h
@@ -61,6 +61,7 @@ struct controller_transaction_t{
     char              *ct_reason;        /* Reason of error (if result != SUCCESS) */
     char              *ct_warning;       /* Warning, first encountered */
     struct timeval     ct_timestamp;     /* Timestamp when entering current state */
+    cxobj             *ct_devdata;       /* Generic device data, eg CS_RPC_GENERIC */
 };
 typedef struct controller_transaction_t controller_transaction;
 
@@ -83,6 +84,7 @@ extern "C" {
 #endif
 
 int   controller_transaction_state_set(controller_transaction *ct, transaction_state state, transaction_result result);
+int   transaction_devdata_add(clixon_handle h, controller_transaction *ct, char *name, cxobj *devdata);
 int   controller_transaction_notify(clixon_handle h, controller_transaction *ct);
 int   controller_transaction_new(clixon_handle h, uint32_t ce_id, char *description, controller_transaction **ct, cbuf **cberr);
 int   controller_transaction_free(clixon_handle h, controller_transaction *ct);